process_items: update project timestamp when necessary.
When project files are uploaded, a directory is created, a file is
renamed, files are moved, or files are deleted, the project timestamp
needs to be updated.

All of these operations are handled by process_items, which is called
by process_files_post, which is called by either project_files or
copyedit_submission.

In each of these cases, update the project timestamp by calling
content_modified() after the ActiveProjectFilesForm does its work
(regardless of whether it returns an error, since some file operations
might have succeeded while others failed.)

If the ActiveProjectFilesForm returns a validation error (is_valid()
is false), then nothing happens, so we do not need to update the
timestamp.

Note that project_files was previously setting modified_datetime after
calling process_files_post - but the project was never saved, so this
line had no effect.  copyedit_submission did not do this.