New function physionet.utility.normalize_zip_permissions.
When we create a zip archive for distribution, the file permissions
should be independent of the permissions that are used on the
PhysioNet server.

(Permissions should be the same regardless of whether the file is "new
in this version of the project" or "identical to a previous version",
and should always be writable by the owner since that is the normal
expectation for Unix programs.)

There is no way to do this using /usr/bin/zip or using the Python
zipfile library (short of either changing file permissions on disk or
somehow spoofing the permissions that the program sees.)

However, we can correct the permissions after generating the zip file,
by patching one four-byte field in each of the central directory
entries.  Parsing the central directory itself is not entirely
trivial, so we delegate that task to /usr/bin/zipdetails (a program
which, surprisingly, is part of the perl package in Debian.)