zip_dir: normalize permissions after writing a ZIP file.
When we create a zip archive for distribution, the file permissions
should be independent of the permissions that are used on the
PhysioNet server.  To do this, after we create the archive in the
usual way (using /usr/bin/zip), invoke normalize_zip_permissions to
correct the permission bits within the archive.