move_files_as_readonly: create ZIP file after setting permissions.
When we create a ZIP archive automatically as part of project
publication, the results should be the same as if the ZIP file is
later created through the console interface.

Furthermore, executable scripts in the ZIP file should be marked as
executable, and other files should be non-executable, regardless of
what permissions the files had when they were uploaded.

Thus, move the "make_zip" step to after setting the published file
permissions.

The license and SHA256SUMS files should still be generated before
setting permissions (they should be marked read-only as all other
files are, and should be included in the total uncompressed storage
size.)