Add merge function for associated emails
Check if both users are not the same.
Check that the arguments are users.

There should be at least 1 associated email to each user account.
If an email is the "is_primary_email", then we set that false.

When all the emails have been merged to the main user, we create
an associated email prepending a UUID to the email and set the
email in the user object of the second user to that new email.