Add merge function for cloud information
Check if both users are not the same.
Check that the arguments are users.

Check if there exist cloud information from the main user,
if it doesnt exists then we create it.

If the second user doesnt have any cloud information we dont
do anything.

If there is cloud information on the second user, we check if
the same field has information in the main user. If the field
is empty then we merge the information.

If there existed any cloud information in the second user, then
set it to empty.