CopyeditForm: create license file when copyediting is finished.
The LICENSE.txt file is automatically generated, but it should be
visible to the authors when they are asked to give their final
approval for the project.  Thus, create the file when the editor
completes copyediting.

(Thus, the file will be updated if copyediting is reopened and further
changes are made.)

The file is also re-created by project.models.move_files_as_readonly,
for the benefit of projects that are in the pipeline at the time this
change is deployed.  Once all such projects have been published (or
reopened), that behavior should be removed, to avoid modifying the
project after it has been approved by the authors.