Add merge function for active/archived/published authors
Get all authorship objects for the second user, then check if
the main user is an author of the same project.

IF the main user is NOT an author, then swap the author object
owner keeping the information intact.

IF the main user IS an author, delete the second’s user author
information.