Add merge function for project DUA signatures
Get all DUA signatures from the second user, then check if the
main user has a DUA signature to the same project.
IF the main user does have a signature, delete the one in the second user.
IF the main user does not have a signature, assign it to the main user.