Add merge function for project authorship invitations
If the invitation is done to the main user, delete it.
If there exist an invitation from the main user to the same email and project, delete the second invitation.
If not the 2 stages above, then change the inviter to the main user.

This should only be called once all project authorship changes are done.