serve_file: escape path properly.
When serving a file using X-Accel-Redirect, the value of the header
should be a valid URL; i.e., any special characters should be replaced
with %XX, which nginx in turn will decode in order to find the actual
file on disk.  In particular, it's important that '%' itself is
replaced with '%25'.  (We don't currently allow files containing '%',
but it's conceivable we might.)

This also avoids crashing in cases where a (buggy?) client requests a
URL containing an escaped control character, such as
'/files/foo/1.0/abc.txt%0D'.

The path passed to serve_file is a real filesystem path (any %XX
escapes in the original URL will have been decoded prior to calling
serve_published_project_file), so we must call urllib.parse.quote to
turn it into a URL path.