Add test cases for serving files with X-Accel-Redirect.
With MEDIA_X_ACCEL_ALIAS defined, serve_published_project_file should
not serve the file contents itself, or even check whether the file
exists; it should simply check that the project exists and the client
is authorized to access it, and generate an X-Accel-Redirect response
telling nginx to serve the requested file from the static/media
directory.

Test this behavior, and test that serve_file will apply the correct
URL encoding.  (Note that %XX escapes in the original path passed to
self.client.get will be decoded and passed to
serve_published_project_file as a Unicode string;
serve_published_project_file will convert that to a filesystem path,
and serve_file will convert that back to a URL, which requires
encoding special characters.)