project.quota: new module.
This module defines the QuotaManager class, which is meant to provide
an interface to a (Linux-like) filesystem quota mechanism, or to
emulate one if such a system is not available.

A QuotaManager provides the following properties:
- block_size: apparent block size of the filesystem
- bytes_used: bytes used by the given account
- inodes_used: inodes used by the current account
- bytes_hard and bytes_soft: hard and soft limits for the number of
  bytes that the account may use
- inodes_hard and inodes_soft: hard and soft limits for the number of
  inodes that the account may use

and the following methods:
- refresh: force the above properties to be refreshed from disk
- set_limits: change the limits that will be enforced
- create_toplevel_directory: create a directory within which quota is
  to be enforced
- check_create_file, check_create_directory, check_delete_file,
  check_delete_directory: simulate creating or deleting a file or
  directory

Currently, the only implementation is DemoQuotaManager, which emulates
a quota system in Python.  It enforces a total size limit and a
maximum number of inodes, but ignores regular files that have more
than one hard link and are older than a specified timestamp.

This rather esoteric metric is meant to allow calculating the
incremental size of an active or published project: the directory
might contain some files that belong to an older published project, as
well as files that are newly uploaded, and we only want the
newly-uploaded files to be counted for quota purposes.