PublishedProject: add incremental_storage_size field.
This field will be used to store the total size of files that are
attributed to this project - i.e., the size of files that were
published as part of this project that were not part of any previous
project.

Calculating this value is not trivial, so it will be saved in the
database when a project is published.