Set incremental_storage_size when publishing a project.
This value should represent the size of files that are new in this
version; we use the project's quota manager to calculate that value.

(This is independent of the project's "main_storage_size", which is
the total download size that will be displayed to visitors, and
includes old files as well as new files.)

This is calculated after generating the SHA256SUMS file (and the
LICENSE file, if needed) so the size of that file should be included
in the total usage.