StorageInfo: add published parameter.
If 'published' is passed to the StorageInfo constructor, this
represents the number of bytes of storage that are consumed by
published versions of a project.  These bytes are assumed to be
counted in 'allowance' and 'used' as well.

This allows StorageInfo to be used to separately report usage of
"existing, published files" and "new files in the current version".

In addition to 'published', the following new properties are defined:
- readable_published: published size as a human-readable string
- p_used_old: percentage of 'allowance' represented by 'published'
- p_used_new: percentage of 'allowance' represented by the difference
  between 'used' and 'published'