ActiveProject.storage_used: include published files in total.
To accurately report the amount of storage used by a project, we need
to include both existing published files, as well as new files
uploaded to the current active project.

The size of the existing published files is calculated by
CoreProject.total_published_size, based on the incremental sizes of
the PublishedProjects (which are saved in the database.)

The size of "newly uploaded" files is determined by
QuotaManager.bytes_used.  Since we are using the DemoQuotaManager
implementation, this means we are counting the sizes of all files
within the project directory that are newer than the project creation
time.