Display published size in file storage progress bar.
In /projects/*/files/ and /console/submitted-projects/*/copyedit/,
there is a "progress bar" showing the fraction of the core project's
available storage that is currently used.

If the project already has published versions, the sizes of those
files will be included in the total (since ActiveProject.storage_used
now reports them).  In this case, display the total published size on
a separate line, and highlight the corresponding fraction of the
progress bar in gray rather than blue.  The color is meant to imply
that the published files are fixed, and only the blue portion of the
bar is possible for the author to modify.