physionet.middleware.maintenance: new module.
In order to provide a clean transition from one server to another, it
is necessary to be able to put the site into an effectively
"read-only" state.  To enforce this, the SystemMaintenanceMiddleware
class can be loaded.

To indicate that the relational database should not be altered, we
define a configuration variable SYSTEM_MAINTENANCE_NO_CHANGES.  This
doesn't mean that the database is actually read-only, but it means
that database changes might not be *persistent*, and thus any action
that would have a persistent effect should be denied.

If SYSTEM_MAINTENANCE_NO_CHANGES is true, the middleware will refuse
all HTTP request methods other than GET, HEAD, OPTIONS, and TRACE,
since those four types of requests should not have any side effects.
Since this may not be desirable for all view, this can be overridden
by using a view decorator (similar to django.middleware.csrf and
django.views.decorators.csrf.)

To indicate that the project filesystem should not be altered, we
define a configuration variable SYSTEM_MAINTENANCE_NO_UPLOAD.  This
limitation will need to be enforced by individual views, by raising a
ServiceUnavailable exception as necessary.

If either the request used a disallowed method, or the view raised
ServiceUnavailable, an error page is returned; if possible, the error
message should be defined by SYSTEM_MAINTENANCE_MESSAGE to inform
visitors about the nature and expected duration of the downtime.