move_files_as_readonly: do not recreate license file.
Previously, when a project was published,
project.models.move_files_as_readonly (invoked asynchronously by
ActiveProject.publish) would generate the LICENSE.txt file, as well as
the SHA256SUMS.txt file and (if enabled) the ZIP archive of the
project.

LICENSE.txt is now generated by console.forms.CopyeditForm.save, so
that the file is visible to the authors at the "approval" stage.
Thus, the file doesn't need to be, and shouldn't be, rewritten when
the project is actually published.