Add configuration for maintenance mode.
SystemMaintenanceMiddleware is added to the list of middleware, so
that requests to make changes can be denied while the system is
undergoing maintenance.

To actually activate maintenance mode requires setting the options
SYSTEM_MAINTENANCE_NO_CHANGES, SYSTEM_MAINTENANCE_NO_UPLOAD, and/or
SYSTEM_MAINTENANCE_MESSAGE, which should be defined in .env.