Add credentialing files for user fixture
On #1004 X amount of credentialing applications were introduced
to facilitate the testings of credential applications.

The applications were all made in the console ignoring the file
upload. Here I also add to the demo information a soft link to the
file. This is the same file we used since the beginning of time.

I also fix the fixture with the file location.