In no-upload mode, prevent editing project files.
When SYSTEM_MAINTENANCE_NO_UPLOAD is set, it should not be possible to
upload (or otherwise modify) active project files.

To enforce this, add a check for SYSTEM_MAINTENANCE_NO_UPLOAD in
process_files_post, which is used by the file manipulation forms both
for authors (project_files) and for editors (copyedit_submission).
This should cause a 503 error when attempting to submit any of the
project file forms (but it should still be possible to load the page.)

To reduce confusion, during no-upload mode, also disable the form
controls in project_files (by setting files_editable false) and
display a warning message.