Allow login in no-changes mode.
If the system is in maintenance mode for a significant length of time
(which should be avoided, but might happen) then visitors should be
able to log in with their existing username and password during that
time, in order to view protected or unpublished projects.

Since the form at /login/ requires a POST request, this requires the
view to be explicitly exempt from the restrictions of
SystemMaintenanceMiddleware.