In no-upload mode, prevent console file changes.
When SYSTEM_MAINTENANCE_NO_UPLOAD is set, it should not be possible to
modify published project files, or to publish an active project (which
requires moving its directory to a new location.)

Prevent changes to project files by disabling the
'make_checksum_file', 'make_zip', and 'deprecate_files' actions in
manage_published_project.

Access to publish_submission is disabled completely to avoid giving
the impression that it might be possible to publish the project during
this time.