In no-upload mode, prevent editing the profile photo.
While SYSTEM_MAINTENANCE_NO_UPLOAD is set, it should not be possible
to upload or delete your profile photo.  However, editing other parts
of the profile form should be allowed.  Thus, raise ServiceUnavailable
only if the form is submitted and includes either an uploaded file or
the 'delete_photo' parameter.