Add a way to add new authors to legacy projects
IF a published project is legacy, then another section will appear
to add and only add new authors.

The order of the authors will be sequential as added.
The first author added will be the submitting author.

At the moment I left the corresponding to False as default.

To add the user object, I took a similar approach to the parent project.