Fixes 400 server error during physionet.mit.edu redirect
When trying to access physionet.mit.edu, a 400 server error is encountered with the current nginx config file. Adding this to ALLOWED_HOSTS solves the problem but the top URL reads physionet.mit.edu which isnt what we intend. Therefore, it appears like a rewrite is necessary to achieve our desired result. Suggestions are welcome though since this is my first time handling the servers on my own.