Dynamically generate database content #1078
Fixes the issue of the list of databases not being dynamically generated. In the past, this database list was generated manually which is difficult to keep up with so a dynamic solution was requested. I first filter each project by their access_policy (open, restricted, credentialed) then sort the projects alphabetically by their lowercase title. Fixes #1078.