Modifies about/database_test page #1078
Modifies the about/database_test page to now display just a single list of all the latest versions of the databases sorted in alphabetical order. This replaces the previous layout which was separated by access policy. Additionally, more information is given on how to search for databases as well as the differences between the access policies for each database. Fixes #1078.