Dynamically generate software content #1092
Fixes the issue of the list of software not being dynamically generated. In the past, this software list was generated manually which was difficult to keep up with so a dynamic solution was requested. I display each software project in alphabetical order by their lowercase title after first introducing the search feature, WFDB format, and LightWAVE. Fixes #1092.