Ignore later contacted refernces
Here I ignore if the button of contact the reference was pushed
for a second or more times leaving the value of the times that
the reference was contacted for the first time.

This is in response of the an email from KP that says the following:

One other thing that I mentioned somewhere I think: the "reference contacted" date should show be the first time that the "contact reference" button is clicked.  If I (accidentally) click it a month later, that date shouldn't change.