user.forms: new class SaferImageField.
This is a wrapper for django.forms.fields.ImageField that does
additional checks to attempt to require the input to be a valid JPEG
or PNG file.

This is not what I would call safe, but it is considerably less awful
than simply taking whatever the client uploads and shoveling it into
PIL.  For example, it will reject EPS files.