ContactForm: disallow sending messages from server's domain.
The "contact" form in /about/ allows sending an email message to
contact@physionet.org.

Require the submitter to specify a "From" address that is not in the
physionet.org domain (or other domains listed in EMAIL_FROM_DOMAINS),
to avoid confusion and so that the message will not be treated as a
"legitimate message from physionet.org".