physionet.wsgi: set process name while handling a request.
If, for some reason, the wsgi handler is taking a long time to handle
a single request, we'd like to be able to see what it's currently
doing, without waiting for the request to finish and then trying to
identify it in the logs.

If we are running under uwsgi, we can set the current process name by
calling uwsgi.setprocname.  (Note that the uwsgi python module is
internal to uwsgi itself; it's not visible elsewhere.)  This can then
be seen using 'ps' or 'top'.

Here, we set the process name to include the current request's path
(PATH_INFO), which should be sufficient to tell us what view functions
might need attention.  We hook into the Django request_started and
request_finished signals so that the process name is set very early
after the request is received, and reverts to "uwsgi (idle)" when the
process is idle.