CSVFileView: handle missing newline at EOF.
When displaying a CSV file, we limit the input to a maximum number of
bytes (MAX_BYTES = 512 * 1024) and a maximum number of rows (MAX_ROWS
= 500).  This is done by wrapping the file in a generator function
that yields one line of text at a time, but also raises an exception
once the maximum number of bytes have been read.

However, if the input file does not end with a line feed, this would
create an infinite loop where the generator would repeatedly try to
read more data (since it hasn't yet seen the "end of the line", but
the size limit hasn't been reached.)

Instead, if there's an incomplete line at the end of the file, treat
that as the last line and try to parse it anyway, but display a
warning.