Added the Known references page
This page will show ALL known references ordered by contact date.
At the moment this is NOT a unique list since more than one
applicant can have the same reference. This list at the moment
will be used to remove the reference was contacted at X date.

This is a "similar" list to what KP had on the previos version of PN.