add grad student category. requested by kp.
When a user is submitted a request for credentialed access, they are asked to self-assign themselves to a category. KP has asked for an additional 'Graduate Student' category to be added. This pull request adds the category with the same form behaviour as 'Student' (i.e. a supervisor is required as a reference). Graduate Student appears between Student and Postdoc (it would have been nice if the sequence 0-7 was spaced to allow insertions within the order).