add simple reporting metrics for credentialing applications
Add a page to the console that displays number of applications, number of applications processed, proportion approved, avg time to decision, avg time to contact reference, and the avg time taken for a reference to respond. also makes minor updates the demo data so applications are displayed over 2 years - 2019 and 2020.