message explaining that an account is required to reset password. closes #1132.
We regularly receive messages from people who believe the 'reset password' form is broken because it does not send them an email. In almost all cases, the issue is that the person does not have an account. Accounts were not migrated from the old platform, so it is understandable that some people believe they have an account when they do not.