convert AuthorInvitation email addresses to lowercase. Ref #1149.
User email addresses are always stored in lowercase. AuthorInvitations allow addresses to be added in upper or lowercase. If an AuthorInvitation is created using upper case characters in the address, the object will not match to a user even if one exists. This patch fixes the issue by converting AuthorInvitation addresses to lowercase, using an approach that appears elsewhere (e.g. see line 67 of clean_email in user/forms.py).