add user admin page to the console.
Currently there is no easy way to view details about a specific user. e.g. Checking which projects a user is associated with requires us to query the db directly from the shell. This change adds an admin page for each user and adds some basic information as a starting point.