allow applications for credentialed access to be paused.
adds a PAUSE_CREDENTIALING flag to the environment file. if the flag is true, we display a message saying that credentialing is paused and remove the link to the application form. this change does not prevent the form being submitted, but it does prevent the most common route to accessing the form.