Creates DB sync between PN live and staging
There is a need to keep a synchronized database on Physionet between the live
site and staging for proper testing.

Here I add a cronjob to be run weekly on Sundays to do a full backup dump and
send it to staging.
Once in staging it will destroy the database and load the database from the
full backup.

A log is kept from the database import.