Metadata.update_internal_links: new function.
When a project changes status from active to published, or published
to active, its base "file" URL and its base "display" URL will both
change.

By default, if the project contains any internal links in the project
description, those links should be updated to point to the project's
new location.  For example, if an active project contains a link
pointing to /project/SHuKI1APLrwWCqxSQnSk/files/RECORDS, then when the
project is published it should be updated to point to
/files/mitbih/1.0.0/RECORDS.

This function performs those transformations, and further performs the
standard cleanups enforced by project.utility.LinkFilter.