Update internal links when an active project is published.
When an active project becomes published, any links pointing to the
active project files are no longer valid, and should be updated to
point to the corresponding location in the published project.  This
includes both links that point to "preview" pages, and links or image
tags that point to "raw" files in the project.