Update internal links when a published project is re-activated.
When a new version of a project is created, any links pointing to the
published project files should, by default, be updated to point to the
new version instead.  This includes both links that point to "preview"
pages, and links or image tags that point to "raw" files in the
project.