Test that project internal links are updated automatically.
When an active project is published, any internal links should be
updated to point to the published project.  When a published project
is re-activated, internal links should be updated to point to the
new active project.

Thus, if we include internal links in the demo project, and repeatedly
publish new versions of the same project, each published version
should end up with internal links pointing to itself.