kvm-pntest-create: set if=virtio.
By default, qemu will use the "ide" interface for virtual disks,
rather than the (faster) "virtio" interface.

Normally it doesn't matter what disk interface we use when installing
the system; however, if the GRUB package later needs to be upgraded,
it will become confused if it can't find the same "physical disk" that
was used to install.

Thus, change kvm-pntest-create to use "if=virtio" to match kvm-pntest.