New script /static/custom/js/clear-autosave.js.
This script will check if a cookie called 'saved_cke_fields' has been
set, and if so, delete autosaved data accordingly.  Views that accept
rich text POSTs will need to set that cookie, and the response will
need to include this script.  If the cookie is not set, this script
does nothing.

(This is done using a cookie rather than conditionally including code
in the HTML response because in many cases, the immediate response to
the POST is to redirect the client to a different page; and yet,
reloading that page should not clear the autosave data a second time.)