Removes previous and next record, replaces with plot selected
Previously, if multiple requests were submitted back to back, then all of them would have to be completed before the plot would render. This prevented users from easily removing a lot of signals or adding a lot of signals as each inter-selection would have to be processed first. This removes that by asking the user to manually plot the desired waveform once they feel satisfied.