Build, test and publish image in Github Actions
Builds the PhysioNet docker image and runs tests on the `dev` branch and PRs against the `dev` branch.
Publishes the image to DockerHub with `latest` tag when the action runs on the `dev` branch.
Can be triggered manually to build, test and optionally publish any branch with a custom tag.
All published images are also tagged with corresponding git commit sha.

To run the tests locally in a similar environment to GitHub Actions you can use:

```sh
$ docker-compose -f docker/docker-compose.test.yml build
$ docker-compose -f docker/docker-compose.test.yml run --rm test
$ docker-compose -f docker/docker-compose.test.yml down
```

To successfully publish the image an environment called `dockerhub` with
secrets `DOCKERHUB_USERNAME` and `DOCKERHUB_TOKEN` should be configured.
The environment should also have protection rules specified (for example
setting `dev` as the deployment branch and adding reviewers for
publishing other branches).