TestMixin: better setup/teardown of static/media directories.
If the static and/or media directories contain read-only
subdirectories (which happens as a result of publishing a project), we
must make those directories writable before trying to delete them or
their contents.

On the other hand, there is no need to change the permissions of
*files* in order to delete them, and in some cases these files may not
be owned by us (or may be symlinks to files not owned by us.)

Furthermore, it's safer to set the directory permissions to 0o700
rather than 0o755 in case one of these directories was not intended to
be world-readable.

In case these directories already existed before running the test
suite, we want to delete them completely as part of TestMixin.setUp.