physionet-build-test.yml: only run flake8 on modified lines.
Running "flake8 --diff" (or "pep8 --diff") on a unified diff will
report errors in the context lines, not only in lines that are
actually modified by the patch.

This unfortunately leads to a cascading effect since fixing one line
of code requires fixing style errors in the surrounding lines, which
in turn brings *more* surrounding lines into view.

Instead, the test should only require fixing style errors in lines
that were modified (i.e., the lines that appear in a "-U0" diff.)