submission_info_card.html: show correct prefix in anonymous URL.
When an unpublished project has an anonymous URL, the URL should be
shown in the submission info displayed in the console.

This includes the views
  console.views.submission_info (via console/submission_info.html)
  console.views.edit_submission (via console/edit_submission.html)
  console.views.copyedit_submission (via console/copyedit_submission.html)
  console.views.awaiting_authors (via console/awaiting_authors.html)
  console.views.publish_submission (via console/publish_submission.html)

To ensure that the "anonymous link" works correctly if someone copies
and pastes the text (rather than copying and pasting the link target),
the text should include the correct protocol prefix and canonical
hostname as determined by notification.get_url_prefix().  This is
already defined as the template parameter 'url_prefix'.