.env: new option BULK_DOWNLOAD_HOSTNAME.
This option can be used to specify an alternate hostname that will be
used for bulk download commands (currently, 'wget'.)  This hostname
will be used *in place of* the server's hostname, and therefore it
should ultimately be another name for the same server, or a proxy or
mirror that is able to serve the same resources.

However, it's useful to have the option of promoting a different
hostname for downloads than the hostname used for interactive
browsing.