published_project: display bulk URL prefix in wget command.
In a published project, the "files" section shows an example wget
command that can be used to download the project.

If BULK_DOWNLOAD_HOSTNAME is defined, we want to use that hostname in
the wget command, rather than the current site hostname.  Do this by
passing a template parameter 'bulk_url_prefix' instead of 'url_prefix'
to avoid confusion.  (Note that 'url_prefix' is not currently used
anywhere else on the published project page, so we can remove the old
template parameter.)