submission_info_card.html: display bulk URL prefix in wget command.
In the unpublished project submission info pages, an example wget
command is shown that can be used to download the project.

If BULK_DOWNLOAD_HOSTNAME is defined, we want to use that hostname in
the wget command, rather than the current site hostname.  Do this by
defining a template parameter 'bulk_url_prefix' to avoid confusion.

The following views use submission_info_card.html:
  console.views.submission_info (via console/submission_info.html)
  console.views.edit_submission (via console/edit_submission.html)
  console.views.copyedit_submission (via console/copyedit_submission.html)
  console.views.awaiting_authors (via console/awaiting_authors.html)
  console.views.publish_submission (via console/publish_submission.html)

Therefore, each of these views must determine the bulk URL prefix and
pass it as a template parameter.

('url_prefix' is not currently used here, but it *ought* to be used in
constructing the "anonymous access" URL; see commit
7ac6b5fd3691fda951f1eb6a06667c0792627b82.)