physionet.test_gcs: run tests only if GCS_HOST is defined.
In order to run these tests, a (dummy) GCS server must be running.
Since this isn't done automatically, these tests should not be run by
default.  Moreover, we shouldn't assume that the server is available
at the hostname "gcs" - that hostname doesn't exist unless you've
taken steps to create it.

Therefore, skip these tests unless the configuration variable GCS_HOST
is set explicitly (which is done in docker/docker-compose.test.yml),
as well as skipping the tests if TEST_GCS_INTEGRATION is false.