project_templatetags: handle unknown future access_policy values.
In access_badge and access_description, we assume that access_policy
is one of the three currently-defined access policy values.  This is
not necessarily true if we are in the process of upgrading the server
to a newer version, and it's better for these functions (which are
purely informational) to gracefully handle unknown input values.