lightwave_project_server: handle sub-database requests.
When reading data from an active project, the lightwave server is
chrooted to the active project directory (so it has no access to files
outside the project, protected or otherwise), and its WFDB path is set
to '/'.

However, from the client's perspective, the project slug is used as
the name of the database, so the client will issue requests like

    ?action=rlist&db=SHuKI1APLrwWCqxSQnSk

These query strings need to be translated to refer to the paths that
are seen by sandboxed-lightwave itself.  Normally the 'db' parameter
will be exactly equal to the project slug, but in some cases (reading
a "sub-record" list) it will be a subdirectory.  Therefore,
'db=SHuKI1APLrwWCqxSQnSk' should be replaced with 'db=.', and
'db=SHuKI1APLrwWCqxSQnSk/foo' should be replaced with 'db=foo'.