register: require a delay after loading and before submitting.
To prevent unintelligent spammers from registering a vast number of
accounts, require the client to actually load the page and then wait
at least 15 seconds before submitting it.  Do this by setting a signed
cookie when the registration page is loaded, and checking the cookie
when the form is submitted.

This requires passing a request object to the RegistrationForm, and
invoking set_response_cookies on the rendered response.  Currently,
this is done for user.register but not for sso.sso_register, on the
assumption that in that case the client has already been authenticated
by the external provider.