RegistrationForm: rate-limit registrations per remote IP address.
To prevent one spammer from trying to register a vast number of
accounts, remember the IP address used when an account is registered
(not activated).  Reject new registration attempts if there have been
100 or more accounts registered from this address in the past 24
hours, or if there have been 20 or more accounts registered from this
address in the past one hour that have not been verified.