user.test_views: test that /register/ requires a delay.
/register/ should now reject simple-minded attempts to register an
account without first loading the page, no less than 15 seconds and no
more than one hour before submitting it.  Test that this works.

Also, fix the assertion logic later in the function.  We don't want to
know whether the return value of filter() is None (which it never is);
we want to know whether the query yields any results.