install-pn-test-server: make .env readable only by www-data.
The file .env contains extremely sensitive information and should not
be world-readable.  It'd be better to have a special group just for
accessing this file (and maybe the rest of /physionet/), but for now,
the simplest thing is to make this file readable only by members of
the www-data group.

This is now done on the staging/production/live servers; do the same
in install-pn-test-server for consistency.