demo-project: make new project version look like the original.
Active project #8 (p7TCIMkltNswuOB9FZH1) was, apparently, created as a
new version of published project #5 (demopsn-1.0).  It has the same
CoreProject and has is_new_version=True and version_order=1.

Therefore, make this project actually look like it was cloned from the
original.  For example, the title has to match.

(Otherwise it is very confusing to log in to the demo system as aewj
and see that you can't create a new version of demopsn because there
is a new version active already, when the active project in question
has a completely different title and content.)