test_urls: new module.
This module provides a generic mechanism for testing URL patterns as
defined by urls.py.  There are a few reasons for this:

- To ensure that production code is actually tested by the test suite,
  we want new URLs and views to be tested by default, even if the test
  is simply "the view can be invoked and doesn't raise an exception".

- We want to be able to collect the HTML (or other output) that is
  produced by each view, and check it for syntax errors.  The easiest
  way to do this is to save the output to ordinary .html files which
  can then be processed by other programs (e.g. the Nu HTML
  validator.)

- We want to be able to easily compare the HTML output before and
  after making a change in git (either a change to a single view, or a
  change in the templates that affects many views across the site);
  this requires a systematic and reproducible process for generating
  HTML files.

This module by itself doesn't have any effect.  In general, in order
to test a URL pattern, we need some parameters (such as a project
slug) that we can plug into it.  Typically there also needs to be some
corresponding example data defined in the demo-*.json files.

Since the goal is for testing to happen automatically when a new
pattern is added to urls.py, and since urls.py is the place where the
parameter names are defined, the test-case parameters are also defined
in urls.py.  They can be defined in two ways:

- A global variable called TEST_DEFAULTS, which is a dictionary
  mapping parameter names to values, and applies to all URLs defined
  in the same 'urls.py' file.

- A global variable called TEST_CASES, which is a dictionary mapping
  the *pattern name* (the string after 'name=' in path()) to a
  dictionary of parameters, or a list of dictionaries of parameters.

The values in TEST_CASES, if any, override those in TEST_DEFAULTS.  If
TEST_DEFAULTS is not set, then only patterns named in TEST_CASES will
be tested.  If TEST_DEFAULTS is set, then all patterns in the module
will be tested, unless TEST_CASES suppresses them.

(In the trivial case where views don't require any parameters, the
module can simply define "TEST_DEFAULTS = {}".)

The intention is that TEST_DEFAULTS should be used in most cases,
which is also meant to gently encourage consistency and clarity in the
naming of parameters.

To export HTML files, set the environment variable
PHYSIONET_TEST_HTML_DIR when running the test suite.  This will save
each response to a file underneath that directory, similar to what
'wget -x -nH' would do.